refactor(mugration): extract domain logic from commands to src/mugration/
- Move execute_mugration, validate_weight_coverage, compute_pi_from_weights, compute_pi_uniform, apply_pseudo_counts, WeightCoverageResult to src/mugration/mugration.rs
- Move MugrationInput, MugrationParams to src/mugration/input.rs
- Move MugrationResult, MugrationGtrOutput, MugrationConfidenceOutput, MugrationTraitsOutput to src/mugration/result.rs
- commands/mugration/ retains only CLI orchestration: args, run (parse + write), comment_provider
- Update all test imports to use crate::mugration::
