+++ packages/treetime/src/commands/mugration/mod.rs
@@ -1,7 +1,5 @@
 pub mod args;
 pub mod comment_provider;
-pub mod input;
-pub mod output;
 pub mod run;
 
 #[cfg(test)]
